requirements.txt
pinecone-client
langchain
python-dotenv
fastapi
uvicorn
python-multipart
pdfplumber
tiktoken
